refactor: Move Parser to top level package
Mistakenly created it in the Task package.
