+++ src/main/java/duke/Duke.java
@@ -1,7 +1,6 @@
 package duke;
 
 import duke.command.Command;
-import duke.task.Parser;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
